+++ apps/ledger-live-desktop/src/renderer/components/DeviceAction/index.tsx
@@ -36,6 +36,7 @@ import {
   renderAllowLanguageInstallation,
   renderInstallingLanguage,
   renderLockedDeviceError,
+  RenderDeviceNotOnboardedError,
 } from "./rendering";
 
 type Props<R, H, P> = {
@@ -271,18 +272,14 @@ export const DeviceActionDefaultRendering = <R, H, P>({
       });
     }
 
-    // NB Until we find a better way, remap the error if it's 6d06 and we haven't fallen
+    // NB Until we find a better way, remap the error if it's 6d06 (LNS, LNSP, LNX) or 6d07 (Stax) and we haven't fallen
     // into another handled case.
     if (
       error instanceof DeviceNotOnboarded ||
-      (error instanceof TransportStatusError && error.message.includes("0x6d06"))
+      (error instanceof TransportStatusError &&
+        (error.message.includes("0x6d06") || error.message.includes("0x6d07")))
     ) {
-      return renderError({
-        t,
-        error: new DeviceNotOnboarded(),
-        withOnboardingCTA: true,
-        info: true,
-      });
+      return <RenderDeviceNotOnboardedError t={t} device={device} />;
     }
 
     if (error instanceof NoSuchAppOnProvider) {