+++ apps/ledger-live-desktop/src/renderer/components/WebPlatformPlayer/WebView.tsx
@@ -1,4 +1,4 @@
-import { shell, WebviewTag } from "electron";
+import { WebviewTag } from "electron";
 import * as remote from "@electron/remote";
 import { JSONRPCRequest } from "json-rpc-2.0";
 import React, { useCallback, useEffect, useMemo, useRef, useState } from "react";
@@ -17,7 +17,7 @@ import {
   signMessageLogic,
 } from "@ledgerhq/live-common/platform/logic";
 import { serializePlatformSignedTransaction } from "@ledgerhq/live-common/platform/serializers";
-import { AppManifest } from "@ledgerhq/live-common/platform/types";
+import { LiveAppManifest } from "@ledgerhq/live-common/platform/types";
 import { useJSONRPCServer } from "@ledgerhq/live-common/platform/JSONRPCServer";
 import {
   RawPlatformSignedTransaction,
@@ -44,7 +44,7 @@ import {
   broadcastTransactionLogic,
   RequestAccountParams,
 } from "./LiveAppSDKLogic";
-import { Container, Wrapper, CustomWebview, Loader } from "./styled";
+import { Container, Wrapper, Loader } from "./styled";
 
 const tracking = trackingWrapper(track);
 
@@ -53,7 +53,7 @@ type WebPlatformPlayerConfig = {
 };
 
 type Props = {
-  manifest: AppManifest;
+  manifest: LiveAppManifest;
   onClose?: () => void;
   inputs?: Record<string, any>;
   config?: WebPlatformPlayerConfig;
@@ -352,31 +352,44 @@ export function WebView({ manifest, onClose, inputs = {}, config }: Props) {
     }
   }, [manifest]);
 
-  const handleNewWindow = useCallback(async e => {
-    const protocol = new URL(e.url).protocol;
-    if (protocol === "http:" || protocol === "https:") {
-      await shell.openExternal(e.url);
+  const handleDomReady = useCallback(() => {
+    const webview = targetRef.current;
+    if (!webview) {
+      return;
     }
+
+    const id = webview.getWebContentsId();
+
+    // cf. https://gist.github.com/codebytere/409738fcb7b774387b5287db2ead2ccb
+    // @ts-expect-error: missing typings for api
+    window.api.openWindow(id);
   }, []);
 
   useEffect(() => {
     const webview = targetRef.current;
 
     if (webview) {
-      // For mysterious reasons, the webpreferences attribute does not
-      // pass through the styled component when added in the JSX.
-      webview.webpreferences = "nativeWindowOpen=no";
-      webview.addEventListener("new-window", handleNewWindow);
       webview.addEventListener("did-finish-load", handleLoad);
+      webview.addEventListener("dom-ready", handleDomReady);
     }
 
     return () => {
       if (webview) {
-        webview.removeEventListener("new-window", handleNewWindow);
         webview.removeEventListener("did-finish-load", handleLoad);
+        webview.removeEventListener("dom-ready", handleDomReady);
       }
     };
-  }, [handleLoad, handleNewWindow]);
+  }, [handleLoad, handleDomReady]);
+
+  const webviewStyle = useMemo(() => {
+    return {
+      opacity: widgetLoaded ? 1 : 0,
+      border: "none",
+      width: "100%",
+      flex: 1,
+      transition: "opacity 200ms ease-out",
+    };
+  }, [widgetLoaded]);
 
   return (
     <Container>
@@ -389,11 +402,25 @@ export function WebView({ manifest, onClose, inputs = {}, config }: Props) {
         config={config?.topBarConfig}
       />
       <Wrapper>
-        <CustomWebview
+        <webview
           src={url.toString()}
           ref={targetRef}
-          style={{ opacity: widgetLoaded ? 1 : 0 }}
+          /**
+           * There seems to be an issue between Electron webview and styled-components
+           * (and React more broadly, cf. comment below).
+           * When using a styled webview component, the `allowpopups` prop does not
+           * seem to be set
+           */
+          style={webviewStyle}
           preload={`file://${remote.app.dirname}/webviewPreloader.bundle.js`}
+          /**
+           * There seems to be an issue between Electron webview and react
+           * Hence, the normal `allowpopups` prop does not work and we need to
+           * explicitly set its value to "true" as a string
+           * cf. https://github.com/electron/electron/issues/6046
+           */
+          // @ts-expect-error: see above comment
+          allowpopups="true"
         />
         {!widgetLoaded ? (
           <Loader>