+++ apps/ledger-live-mobile/src/components/RootNavigator/types/BaseNavigator.ts
@@ -50,6 +50,7 @@ import type { WalletConnectNavigatorParamList } from "./WalletConnectNavigator";
 import type { WalletConnectLiveAppNavigatorParamList } from "./WalletConnectLiveAppNavigator";
 import type { PostOnboardingNavigatorParamList } from "./PostOnboardingNavigator";
 import type { CustomImageNavigatorParamList } from "./CustomImageNavigator";
+import type { ClaimNftNavigatorParamList } from "./ClaimNftNavigator";
 import type { PolkadotSimpleOperationFlowParamList } from "../../../families/polkadot/SimpleOperationFlow/types";
 import type { PolkadotNominateFlowParamList } from "../../../families/polkadot/NominateFlow/types";
 import type { PolkadotUnbondFlowParamList } from "../../../families/polkadot/UnbondFlow/type";
@@ -120,6 +121,9 @@ export type BaseNavigatorStackParamList = {
     defaultTicker?: string;
   };
   [ScreenName.Learn]: undefined;
+  [ScreenName.LearnWebView]: {
+    uri?: string;
+  };
   [ScreenName.SwapOperationDetails]: {
     swapOperation: MappedSwapOperation;
   };
@@ -267,6 +271,7 @@ export type BaseNavigatorStackParamList = {
     WalletConnectNavigatorParamList | WalletConnectLiveAppNavigatorParamList
   >;
   [NavigatorName.CustomImage]: NavigatorScreenParams<CustomImageNavigatorParamList>;
+  [NavigatorName.ClaimNft]: NavigatorScreenParams<ClaimNftNavigatorParamList>;
   [NavigatorName.PostOnboarding]: NavigatorScreenParams<PostOnboardingNavigatorParamList>;
 
   // Polkadot