+++ apps/ledger-live-mobile/src/components/WebPlatformPlayer/WebViewV2.tsx
@@ -27,6 +27,7 @@ import { getEnv } from "@ledgerhq/live-common/env";
 import type { Transaction } from "@ledgerhq/live-common/generated/types";
 import type { AppManifest } from "@ledgerhq/live-common/wallet-api/types";
 import {
+  bitcoinFamillyAccountGetXPubLogic,
   broadcastTransactionLogic,
   receiveOnAccountLogic,
   signTransactionLogic,
@@ -150,7 +151,7 @@ export const WebView = ({ manifest, inputs }: Props) => {
       };
       serverRef.current = new WalletAPIServer(transportRef.current);
       serverRef.current.setPermissions({
-        currencyIds: manifest.currencies === "*" ? ["*"] : manifest.currencies,
+        currencyIds: manifest.currencies === "*" ? ["**"] : manifest.currencies,
         methodIds: manifest.permissions as unknown as string[], // TODO use the new manifest type for v2 as we should avoid as typings
       });
       serverRef.current.setAccounts(walletAPIAccounts);
@@ -527,6 +528,13 @@ export const WebView = ({ manifest, inputs }: Props) => {
         tracking.deviceCloseSuccess(manifest);
 
         return Promise.resolve(transportId);
+      });
+
+      serverRef.current.setHandler("bitcoin.getXPub", ({ accountId }) => {
+        return bitcoinFamillyAccountGetXPubLogic(
+          { manifest, accounts, tracking },
+          accountId,
+        );
       });
     }
     // Only used to init the server, no update needed