+++ apps/ledger-live-mobile/src/screens/Swap/Form/Connect.tsx
@@ -2,8 +2,10 @@ import React, { useState, useCallback } from "react";
 import { Flex } from "@ledgerhq/native-ui";
 import connectManager from "@ledgerhq/live-common/hw/connectManager";
 import { createAction } from "@ledgerhq/live-common/hw/actions/manager";
+import { useFeature } from "@ledgerhq/live-common/featureFlags/index";
 import { Device } from "@ledgerhq/live-common/hw/actions/types";
 import SelectDevice from "../../../components/SelectDevice";
+import SelectDevice2 from "../../../components/SelectDevice2";
 import DeviceActionModal from "../../../components/DeviceActionModal";
 import { TrackScreen } from "../../../analytics";
 import SkipSelectDevice from "../../SkipSelectDevice";
@@ -21,6 +23,8 @@ export function Connect({
   const [device, setDevice] = useState<Device>();
   const [result] = useState();
 
+  const newDeviceSelectionFeatureFlag = useFeature("llmNewDeviceSelection");
+
   const onModalHide = useCallback(() => {
     if (result) {
       // Nb need this in order to wait for the first modal to hide
@@ -37,7 +41,13 @@ export function Connect({
         provider={provider}
       />
       <SkipSelectDevice onResult={setDevice} />
-      <SelectDevice onSelect={setDevice} autoSelectOnAdd />
+      {newDeviceSelectionFeatureFlag?.enabled ? (
+        <Flex px={16} py={8} flex={1}>
+          <SelectDevice2 onSelect={setDevice} stopBleScanning={!!device} />
+        </Flex>
+      ) : (
+        <SelectDevice onSelect={setDevice} autoSelectOnAdd />
+      )}
       <DeviceActionModal
         onClose={() => setDevice(undefined)}
         onModalHide={onModalHide}